requirements.txt
aiogram~=3.13.1
pydantic-settings~=2.6.1
redis
openai
pytz
httpx
groq~=0.12.0
loguru~=0.6.0
gigachat~=0.1.36
langchain-core~=0.3.19
python-dotenv~=1.0.1
peewee~=3.17.8